if file exists remove it
Former-commit-id: 4d8b366643681fb3ea330499bd109c0c4220826f

+++ install_generic_tech.py
@@ -21,7 +21,8 @@ def install_generic_tech(src, dest):
 
     dest_folder = dest.parent
     dest_folder.mkdir(exist_ok=True, parents=True)
-    os.remove(dest)
+    if os.path.isfile(dest):
+        os.remove(dest)
     os.symlink(src, dest)
     print(f"generic layermap installed to {dest}")
